refactor: move replicate lib
